add a new type of module: vlc_renderer
This new type of module will be used by renderers (chromecast, UPnP Renderer,
miracast, airport, DIAL, ConeCast) to configure a VLC input_thread for
rendering, like configuring a sout according to the renderer ip and
capabilities.

This API is experimental and has been tested only for chromecast, it may move
in the future when implementing others renderers.

There is an unique vlc_renderer object attached to the libvlc instance, that is
initialized from libvlc_InternalInit() (only the object is allocated, a module
is loaded only if specified from command line).

Also-by: Steve Lhomme <robux4@videolabs.io>